+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## tonic, etcd [0.3.0] - 2023-07-20
+
+### Changed
+
+- tonic: Update `tonic` and `tonic-build` to v0.9.2.
+- etcd: Update `etcd-client` to v0.11.1.
+
 ## [0.2.25] - 2023-07-03
 
 ### Added